1.38.d11 warning修正(commit test2)
git-svn-id: svn+ssh://svn.osdn.jp/svnroot/saccubus/branches/dev_branch@424 c31a3d4d-7098-419a-a5c4-0623780f6bbf

+++ frontend/src/saccubus/Converter.java
@@ -1578,11 +1578,11 @@ public class Converter extends Thread {
 			int height = videoAspect.getHeight();
 			double rate;
 			if(videoAspect.isWide()){
-				rate = 640.0 / (double)width;
+				rate = 640.0 / width;
 				width = 640;
 				height *= rate;
 			}else{
-				rate = 384.0 / (double)height;
+				rate = 384.0 / height;
 				width *= rate;
 				height = 384;
 			}
